Create warehouse_aggregation
Aggregate the data from two tables to a table with applying subqueries. Objectives of the project: Is to aggregate the data into a table containing each "Warehouse_id," "State," "Alias," and "Number of orders" (total orders for all warehouses). A column
classifies each warehouse by the percentage of total orders that it: "Fulfilled 0-20%, 21-60%, or >60%".